% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{analyze_cytof_file}
\alias{analyze_cytof_file}
\title{Analyze a single CyTOF file}
\usage{
analyze_cytof_file(file, dir_in, dir_out, cols, cofactor = 5)
}
\arguments{
\item{file}{FCS file name.}

\item{dir_in}{A path to read fcs files from.}

\item{dir_out}{A path to write output.}

\item{cols}{Columns to use in analysis.}

\item{cofactor}{Parameter for asinh transformation.}
}
\value{
Nothing; all output written to file.
}
\description{
Read and parse file, cluster and label main cell types, detect
debris and doublets, gate downstream T and B cell subpopulations.
}
