% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_dirs}
\alias{create_dirs}
\title{Create output directory structure}
\usage{
create_dirs(dir_out)
}
\arguments{
\item{dir_out}{Base output path.}
}
\value{
Nothing.
}
\description{
Create subdirectories for saving figures and csv output.
}
