% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{analyze_cytof_file}
\alias{analyze_cytof_file}
\title{Analyze a single CyTOF file}
\usage{
analyze_cytof_file(file, dir_in, dir_out, cols, defs, defs_tcell, defs_myeloid)
}
\arguments{
\item{file}{FCS file name.}

\item{dir_in}{A path to read fcs files from.}

\item{dir_out}{A path to write output.}

\item{cols}{Columns to use in analysis.}

\item{defs}{Model definitions for main cell types.}

\item{defs_tcell}{Model definitions for T cell subtypes.}

\item{defs_myeloid}{Model definitions for monocytes and DCs.}
}
\value{
Nothing; all output written to file.
}
\description{
Read and parse file, cluster and label main cell types, detect
debris and doublets, gate downstream T and B cell subpopulations.
}
